requirements.txt
aiogram==3.12.0
  - aiofiles [required: ~=23.2.1, installed: 23.2.1]
  - aiohttp [required: >=3.9.0,<3.11, installed: 3.10.5]
    - aiohappyeyeballs [required: >=2.3.0, installed: 2.4.0]
    - aiosignal [required: >=1.1.2, installed: 1.3.1]
      - frozenlist [required: >=1.1.0, installed: 1.4.1]
    - async-timeout [required: >=4.0,<5.0, installed: 4.0.3]
    - attrs [required: >=17.3.0, installed: 24.2.0]
    - frozenlist [required: >=1.1.1, installed: 1.4.1]
    - multidict [required: >=4.5,<7.0, installed: 6.0.5]
    - yarl [required: >=1.0,<2.0, installed: 1.9.4]
      - idna [required: >=2.0, installed: 3.7]
      - multidict [required: >=4.0, installed: 6.0.5]
  - certifi [required: >=2023.7.22, installed: 2024.7.4]
  - magic-filter [required: >=1.0.12,<1.1, installed: 1.0.12]
  - pydantic [required: >=2.4.1,<2.9, installed: 2.8.2]
    - annotated-types [required: >=0.4.0, installed: 0.7.0]
    - pydantic_core [required: ==2.20.1, installed: 2.20.1]
      - typing_extensions [required: >=4.6.0,!=4.7.0, installed: 4.12.2]
    - typing_extensions [required: >=4.6.1, installed: 4.12.2]
  - typing_extensions [required: >=4.7.0,<=5.0, installed: 4.12.2]
Jinja2==3.1.4
  - MarkupSafe [required: >=2.0, installed: 2.1.5]
loguru==0.7.2
  - colorama [required: >=0.3.4, installed: 0.4.6]
  - win32-setctime [required: >=1.0.0, installed: 1.1.0]
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
  - certifi [required: >=2017.4.17, installed: 2024.7.4]
  - charset-normalizer [required: >=2,<4, installed: 3.3.2]
  - idna [required: >=2.5,<4, installed: 3.7]
  - urllib3 [required: >=1.21.1,<3, installed: 2.2.2]
setuptools==73.0.1
